% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFMfunction.R
\name{RFMfunction}
\alias{RFMfunction}
\title{calculate RFM}
\usage{
RFMfunction(data, weight_recency = 1, weight_frequency = 1,
  weight_monetary = 1)
}
\arguments{
\item{data}{- A data.table containing the transaction record details for every customer}

\item{weight_recency}{- weight of recency}

\item{weight_frequency}{- wegiht of frequency}

\item{weight_monetary-}{weight of monetary value}
}
\value{
Returns a data.table containing the recency, frequency and monetary scores as well as the weighted final score and the group membership.
}
\description{
Calculated the weighted RFM score: recency, frequency, monetary value for customers
}
\details{
\code{data} contains the transactional data. The dataset must contain a column labeled "Customer"
that allows unique customer identification and a column called "TransDate", indicating the purchase date.
The column "PurchAmount" specifies the total spendig of the purchase.
}
\examples{

some example
dataResult <- RFMfunction(salesClients, 20, 20, 60)
some other example

}
